ch(endpoints):adding environment variables -created and configured .env -modified the controllers, to remove all codes for validation -modified middlewares, to ensure DRY -modified the routes, due to the change in middlewares [finishes #170009189]
