+++ .gitignore
@@ -2,3 +2,5 @@ node_modules
 coverage
 .nyc_output
 package-lock.json
+build 
+.env
