--- .nyc_output/processinfo/index.json
@@ -1,1 +1,0 @@
-{"processes":{"d07ff308-19fb-43cc-b739-412c9f90ec73":{"parent":null,"children":[]}},"files":{"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\server.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\routes\\auth.route.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\controllers\\users.controller.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\models\\user.model.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\models\\data.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\models\\redflag.model.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\middlewares\\auth.middleware.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\routes\\redflags.route.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\controllers\\redflags.controller.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"],"E:\\roger\\school materials\\Andela\\our bootcamp\\Broadcaster\\server\\middlewares\\token.middleware.js":["d07ff308-19fb-43cc-b739-412c9f90ec73"]},"externalIds":{}}
